# Improve ext2 (again), primarily around directory entry