This PR primarily revises the implementation of ext2's directory entry.
The key observation is that we rarely access the "name" (256 byte `FixedCstr`) of a direntry, but its header mostly. So we only touch the "name" field when necessary while iterating through directory entries.

Below shows the benefit on benchmark `sqlite-speedtest1-ext2`:

| Case                 | Old (latency) | New     |
| -------------------- | ------------- | ------- |
| SELECTS on an IPK    | 5.456s        | 4.407s  |
| SELECTS on a TEXT PK | 5.821s        | 4.757s  |
| ...                  |               |         |
| TOTAL                | 83.059s       | 77.178s |