@lucassong-mh 2024-12-27
> Any data on the performance improvements?

Sorry I've forgot to. The improvements have been updated above.